1.21.9~1.21.10

+++ src/client/java/lpctools/lpcfymasaapi/screen/ChooseItemScreen.java
@@ -108,7 +108,8 @@ public class ChooseItemScreen extends GuiBase {
 			y += h;
 		}
 	}
-	@Override public boolean onMouseScrolled(int mouseX, int mouseY, double horizontalAmount, double verticalAmount) {
+	
+	@Override public boolean onMouseScrolled(double mouseX, double mouseY, double horizontalAmount, double verticalAmount) {
 		int lastShift = shift;
 		shift -= (int)Math.signum(verticalAmount);
 		int mh = ((searchedItems.size() + width - 1) / width) - height;
